build-sys: create static uuid_generate_*.3 files
It seems more simple to create small three files with man page links
than generate the files on the fly.

Note that the libuuid/man directory is not in builddir by default so
you have to create proper directory and then create the man pages...
too complex task for this three tiny files.

Signed-off-by: Karel Zak <kzak@redhat.com>